+++ src/LinkDotNet.StringBuilder/ValueStringBuilder.cs
@@ -248,52 +248,57 @@ public ref partial struct ValueStringBuilder : IDisposable
     /// Returns the index within this string of the first occurrence of the specified substring.
     /// </summary>
     /// <param name="word">Word to look for in this string.</param>
+    /// <param name="comparisonType">One of the enumeration values that specifies the rules for the search.</param>
     /// <returns>The index of the found <paramref name="word"/> in this string or -1 if not found.</returns>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public readonly int IndexOf(ReadOnlySpan<char> word) => IndexOf(word, 0);
+    public readonly int IndexOf(ReadOnlySpan<char> word, StringComparison comparisonType = StringComparison.Ordinal) => IndexOf(word, 0, comparisonType);
 
     /// <summary>
     /// Returns the index within this string of the first occurrence of the specified substring, starting at the specified index.
     /// </summary>
     /// <param name="word">Word to look for in this string.</param>
     /// <param name="startIndex">Index to begin with.</param>
+    /// <param name="comparisonType">One of the enumeration values that specifies the rules for the search.</param>
     /// <returns>The index of the found <paramref name="word"/> in this string or -1 if not found.</returns>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public readonly int IndexOf(ReadOnlySpan<char> word, int startIndex)
+    public readonly int IndexOf(ReadOnlySpan<char> word, int startIndex, StringComparison comparisonType = StringComparison.Ordinal)
     {
-        return buffer[startIndex..bufferPosition].IndexOf(word);
+        return buffer[startIndex..bufferPosition].IndexOf(word, comparisonType);
     }
 
     /// <summary>
     /// Returns the index within this string of the last occurrence of the specified substring.
     /// </summary>
     /// <param name="word">Word to look for in this string.</param>
+    /// <param name="comparisonType">One of the enumeration values that specifies the rules for the search.</param>
     /// <returns>The index of the found <paramref name="word"/> in this string or -1 if not found.</returns>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public readonly int LastIndexOf(ReadOnlySpan<char> word) => LastIndexOf(word, 0);
+    public readonly int LastIndexOf(ReadOnlySpan<char> word, StringComparison comparisonType = StringComparison.Ordinal) => LastIndexOf(word, 0, comparisonType);
 
     /// <summary>
     /// Returns the index within this string of the last occurrence of the specified substring, starting at the specified index.
     /// </summary>
     /// <param name="word">Word to look for in this string.</param>
     /// <param name="startIndex">Index to begin with.</param>
+    /// <param name="comparisonType">One of the enumeration values that specifies the rules for the search.</param>
     /// <returns>The index of the found <paramref name="word"/> in this string or -1 if not found.</returns>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public readonly int LastIndexOf(ReadOnlySpan<char> word, int startIndex)
+    public readonly int LastIndexOf(ReadOnlySpan<char> word, int startIndex, StringComparison comparisonType = StringComparison.Ordinal)
     {
-        return buffer[startIndex..bufferPosition].LastIndexOf(word);
+        return buffer[startIndex..bufferPosition].LastIndexOf(word, comparisonType);
     }
 
     /// <summary>
     /// Returns whether a specified substring occurs within this string.
     /// </summary>
     /// <param name="word">Word to look for in this string.</param>
+    /// <param name="comparisonType">One of the enumeration values that specifies the rules for the search.</param>
     /// <returns>True if the value parameter occurs within this string, or if value is the empty string (""); otherwise, false.</returns>
     /// <remarks>
     /// This method performs an ordinal (case-sensitive and culture-insensitive) comparison.
     /// </remarks>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public readonly bool Contains(ReadOnlySpan<char> word) => IndexOf(word) != -1;
+    public readonly bool Contains(ReadOnlySpan<char> word, StringComparison comparisonType = StringComparison.Ordinal) => IndexOf(word, comparisonType) != -1;
 
     /// <summary>
     /// Returns whether the characters in this builder are equal to the characters in the given span.